Coding Assignments/annotation-format-converter-redbrick-ai/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
Pillow==8.3.1
pydantic==1.8.2
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
